requirements.txt
ocrmypdf==14.0.1
Pillow==9.2.0
requests==2.25.1
